+++ CHANGELOG.md
@@ -7,7 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.37.0](https://github.com/Substra/substra/releases/tag/0.37.0) - 2022-09-19
+
+### Fixed
+
+tests: fix algo inputs and task code consistency
+
 ### Changed
+
 - BREAKING CHANGE: `schemas.ComputePlanSpec` `clean_models` property is now removed, the transient property on tasks outputs should be used instead.
 
 ## [0.36.0](https://github.com/Substra/substra/releases/tag/0.36.0) - 2022-09-12
@@ -61,7 +68,6 @@ client.list_composite_traintuple(filters={"status": [substra.models.Status.doing
 ### Documentation
 
 - Add the filters documentation to the references
-
 
 ## [0.35.0](https://github.com/Substra/substra/releases/tag/0.35.0) - 2022-08-29